requirements.txt
Necessary dependencies

pip install the following:

nltk
gensim.models
torch
rake_nltk
summa